finalized commit

+++ ssp_assignment/app/Http/Controllers/CategoryController.php
@@ -6,20 +6,31 @@ use App\Http\Requests\CategoryRequest;
 use App\Models\Category;
 use App\Models\Product;
 use Illuminate\Http\Request;
+use Illuminate\Support\Facades\Auth;
 
 class CategoryController extends Controller
 {
     public function index()
     {
-        return redirect()->route('admin.category.index');
+        if (Auth::check() && auth()->user()->role->value == 1) {
+
+            return redirect()->route('admin.category.index');
+        } else {
+            abort(403, 'Unauthorized Access');
+        }
     }
 
     public function create()
     {
-        return view('windmill-admin.category.form', [
-            'category' => new Category(),
-            'purpose' => 'Create Category',
-        ]);
+        if (Auth::check() && auth()->user()->role->value == 1) {
+
+            return view('windmill-admin.category.form', [
+                'category' => new Category(),
+                'purpose' => 'Create Category',
+            ]);
+        } else {
+            abort(403, 'Unauthorized Access');
+        }
     }
 
     public function store(Request $request)
@@ -43,10 +54,15 @@ class CategoryController extends Controller
 
     public function edit(Category $category)
     {
-        return view('windmill-admin.category.form', [
-            'category' => $category,
-            'purpose' => 'Edit Category',
-        ]);
+        if (Auth::check() && auth()->user()->role->value == 1) {
+
+            return view('windmill-admin.category.form', [
+                'category' => $category,
+                'purpose' => 'Edit Category',
+            ]);
+        } else {
+            abort(403, 'Unauthorized Access');
+        }
     }
 
     public function update(Request $request, Category $category)
@@ -60,8 +76,13 @@ class CategoryController extends Controller
 
     public function destroy(Category $category)
     {
-        $category->delete();
+        if (Auth::check() && auth()->user()->role->value == 1) {
 
-        return redirect()->route('admin.category.index')->with('success', 'Category Deleted Successfully!');
+            $category->delete();
+
+            return redirect()->route('admin.category.index')->with('success', 'Category Deleted Successfully!');
+        } else {
+            abort(403, 'Unauthorized Access');
+        }
     }
 }
